Refactor paredit configuration caching
Move paredit configuration caching logic to the extension module.
This change centralizes the configuration handling and ensures that
the cached configuration is properly invalidated on changes,
improving performance and maintainability.
