+++ src/paredit/extension.ts
@@ -15,9 +15,11 @@ import * as docMirror from '../doc-mirror/index';
 import { EditableDocument } from '../cursor-doc/model';
 import { assertIsDefined } from '../utilities';
 import * as config from '../formatter-config';
+import * as mainConfig from '../config';
 import * as textNotation from '../extension-test/unit/common/text-notation';
 import * as calvaState from '../state';
-import { defaultBindingForms, getPareditConfig } from '../cursor-doc/paredit-config';
+import { createPareditConfig, defaultBindingForms } from '../cursor-doc/paredit-config';
+import type { PareditConfig } from '../cursor-doc/paredit-config';
 
 const onPareditKeyMapChangedEmitter = new EventEmitter<string>();
 
@@ -44,6 +46,21 @@ function shouldKillAlsoCutToClipboard(override?: boolean): boolean {
 
 function multiCursorEnabled(override?: boolean): boolean {
   return override ?? workspace.getConfiguration().get('calva.paredit.multicursor');
+}
+
+// Cache paredit configuration
+let pareditConfigCache: PareditConfig | null = null;
+
+/**
+ * Gets the merged paredit configuration from VS Code settings and .calva/config.edn.
+ * Results are cached and invalidated on config changes.
+ */
+export function getPareditConfig(): PareditConfig {
+  if (!pareditConfigCache) {
+    const cfg = mainConfig.getConfig();
+    pareditConfigCache = createPareditConfig(cfg.customPairForms, cfg.customThreadingMacros);
+  }
+  return pareditConfigCache;
 }
 
 type PareditCommand = {
